Add files via upload
This repository contains a collection of JavaScript practice programs that I have created to enhance my coding skills. The programs cover various concepts such as  loops, functions, arrays and objects. These exercise are part of my ongoing journey to improve my proficiency in JavaScript programming. More files will be added to this repository in the future, so stay tuned for updates. 

Feel free to explore the code and provide feedback or suggestions.