Add files via upload
This project is to develop a front-end model of a web-based platform using HTML, CSS, and JavaScript that allows residents to easily submit city complaints and provide feedback. 
This platform aims to offer an intuitive and accessible interface for reporting issues such as potholes, broken streetlights, and other public concerns, as well as sharing general feedback. 
While this project currently focuses on the front-end development, future enhancements may include back-end integration for comprehensive functionality.